requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
google-generativeai==0.3.1
groq==0.4.1
PyPDF2==3.0.1
pillow==10.1.0
requests==2.31.0
